#AOC-531[Modification in user creation according to the CSR code implementation]

+++ src/com/avery/storage/dao/impl/SystemCsrCodeDao.java
@@ -7,5 +7,5 @@ import com.avery.storage.entities.SystemCsrCode;
 
 public interface SystemCsrCodeDao extends GenericDao<SystemCsrCode, Long>{
 
-	public Map<String, Object> getBySystemAndOrgCodeId(int systemId, int orgId);
+	public Map<String, Object> getBySystemAndOrgCodeId(long systemId, long orgId);
 }
